requirements.txt
attrs==21.2.0
aws-cdk.assets==1.127.0
aws-cdk.aws-applicationautoscaling==1.127.0
aws-cdk.aws-autoscaling-common==1.127.0
aws-cdk.aws-cloudformation==1.127.0
aws-cdk.aws-cloudwatch==1.127.0
aws-cdk.aws-codeguruprofiler==1.127.0
aws-cdk.aws-codestarnotifications==1.127.0
aws-cdk.aws-ec2==1.127.0
aws-cdk.aws-ecr==1.127.0
aws-cdk.aws-ecr-assets==1.127.0
aws-cdk.aws-efs==1.127.0
aws-cdk.aws-events==1.127.0
aws-cdk.aws-iam==1.127.0
aws-cdk.aws-imagebuilder==1.127.0
aws-cdk.aws-kms==1.127.0
aws-cdk.aws-lambda==1.127.0
aws-cdk.aws-logs==1.127.0
aws-cdk.aws-s3==1.127.0
aws-cdk.aws-s3-assets==1.127.0
aws-cdk.aws-signer==1.127.0
aws-cdk.aws-sns==1.127.0
aws-cdk.aws-sqs==1.127.0
aws-cdk.aws-ssm==1.127.0
aws-cdk.cloud-assembly-schema==1.127.0
aws-cdk.core==1.127.0
aws-cdk.custom-resources==1.127.0
aws-cdk.cx-api==1.127.0
aws-cdk.region-info==1.127.0
cattrs==1.8.0
cdk-expects-matcher==0.1.2
constructs==3.3.161
expects==0.9.0
gitdb==4.0.7
GitPython==3.1.24
iniconfig==1.1.1
jsii==1.38.0
packaging==21.0
pluggy==1.0.0
publication==0.0.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
smmap==4.0.0
toml==0.10.2
typing-extensions==3.10.0.2
